==> code/sandbox/06c_try_lioness.Rmd <==
---
title: "06c_try_lioness"
author: "E Flynn"
date: "9/6/2020"
output: html_document
---

Code for trying lioness. Also includes attempt to try jive / start of tSNR.
Wrote several functions for running this. Set up so we get the same sets of genes/edges for a pair of DE comparisons to allow for visualization (make sure to set the same seed!).

Note: coexpression not TF networks in this version.

TODO: null dist or other assessment

Other ideas:
- try std network techniques?
- GSEA w MSigDB, GO analysis
  - immunologic gene signatures "ImmuneSigDB"
  - cancer genes?
- PROPS, STAMS

```{r}
library(lionessR)
library(igraph)
library(limma)
library(SummarizedExperiment)
library(tidyverse)
```

Start with a small dataset
```{r}
study1 <- prev_data %>% filter(first_study=="GSE11784") %>% select(geo_accession, smok, metadata_sex, expr_sex, copd) 
study1.2 <- study1 %>% 
  filter(expr_sex!="unknown", copd == "n",
         (metadata_sex=="M" & expr_sex=="male") |
           (metadata_sex=="F" & expr_sex=="female"))

# get GEO data, normalize + convert to genes
#study1_gse <- getGEO("GSE11784")
gse11784 <- read_tsv("~/Downloads/55b8f63a-b30a-4723-8475-9c5ae629f5a8/GSE11784/GSE11784.tsv")


# convert genes - note slow! #
require('biomaRt')
# ensembl = useMart("ensembl",dataset="hsapiens_gene_ensembl")
# gene_convert =getBM(attributes=c("ensembl_gene_id", "hgnc_symbol"),
#       filters="ensembl_gene_id",
#       values=gse11784$Gene,
#       mart=ensembl)
gene_convert %>% write_csv("data/hgnc_ensembl.csv")
# convert and summarize to mean
gse11784.2 <- gse11784 %>% 
  inner_join(gene_convert, by=c("Gene"="ensembl_gene_id")) %>%

  filter(hgnc_symbol != "") %>%
  dplyr::select(-Gene) %>% 
  mutate(across(-hgnc_symbol, ~log2(.+0.5))) %>%
  mutate(across(-hgnc_symbol, ~replace_na(., replace=0)))

min_val <- min(gse11784.2 %>% dplyr::select(-hgnc_symbol), na.rm=T)
gse11784.3 <-gse11784.2 %>% mutate(across(-hgnc_symbol, ~.+min_val))


study1.3 <- study1.2 %>%
  mutate(is_male=case_when(expr_sex=="female"~"N",
                           expr_sex=="male"~"Y"),
         is_smoker=case_when(smok=="S" ~ "Y",
                             smok=="NS" ~ "N"))

f_meta <- study1.3 %>% filter(expr_sex=="female")
m_meta <- study1.3 %>% filter(expr_sex=="male")


f_dat <- gse11784.3[, f_meta %>% pull(geo_accession)]
rownames(f_dat) <- gse11784.3$hgnc_symbol
m_dat <- gse11784.3[, m_meta %>% pull(geo_accession)]
rownames(m_dat) <- gse11784.3$hgnc_symbol

# separate into males and females
# targets = smokers vs nonsmokers
# ALTERNATE - could look at m vs f smokers

s_meta <- study1.3 %>% filter(smok=="S")
ns_meta <- study1.3 %>% filter(smok=="NS")
s_dat <- gse11784.3[, s_meta %>% pull(geo_accession)]
rownames(s_dat) <- gse11784.3$hgnc_symbol
ns_dat <- gse11784.3[, ns_meta %>% pull(geo_accession)]
rownames(ns_dat) <- gse11784.3$hgnc_symbol

```

PVCA
```{r}
rownames(study1.3) <- study1.3$geo_accession
eset <- ExpressionSet(assayData=as.matrix(gse11784.3[,study1.3$geo_accession]), phenoData=AnnotatedDataFrame(study1.3 %>% dplyr::rename(sex=expr_sex)))
pvcaObj <- pvcaBatchAssess(eset, c("smok","sex"), 0.6)

bp <- barplot(pvcaObj$dat, xlab = "Effects", ylab = "Weighted average proportion variance", ylim= c(0,1.1),col = c("blue"), las=2, main="PVCA estimation bar chart")
axis(1, at = bp, labels = pvcaObj$label, xlab = "Effects", cex.axis = 0.5, las=2)
values = pvcaObj$dat 
new_values = round(values , 3)
text(bp,pvcaObj$dat,labels = new_values, pos=3, cex = 0.8) 

# aov
aov_in <- gse11784.3[,c("hgnc_symbol", study1.3$geo_accession)] %>%
  group_by(hgnc_symbol) %>%
  pivot_longer(-hgnc_symbol, names_to="geo_accession", values_to="expression") %>%
  ungroup() %>%
  dplyr::rename(gene=hgnc_symbol) %>%
  left_join(study1.3 %>% dplyr::select(geo_accession, smok, expr_sex), by="geo_accession") %>%
  dplyr::rename(sex=expr_sex)

aov(gene~mok+sex+smok*sex, data=aov_in)
# VCA

# tSNR

```


```{r}
# make it a summarized EXP
makeSumExp <- function(exp_data, phe_data){
  rowData <- DataFrame(row.names = rownames(exp_data), 
                       gene = rownames(exp_data))
  colData <- phe_data %>% dplyr::rename("sample"="geo_accession")
  rownames(colData) <- colData$sample

  se <- SummarizedExperiment(assays = list(counts = as.matrix(exp_data)), colData = colData, rowData = rowData)
  return(se)
}


# grab 500 most variably expressed genes
varExp <- function(se, nsel=500) {
  cvar <- apply(assay(se), 1, sd)
  dat <- se[tail(order(cvar), nsel), ]
  return(dat)
}

varExpEither <- function(se1, se2, nsel=500){
  cvar1 <- apply(assay(se1), 1, sd)
  cvar2 <- apply(assay(se2), 1, sd)
  l1 <- tail(names(cvar1)[order(cvar1)], nsel)
  l2 <- tail(names(cvar2)[order(cvar2)], nsel)
  genes <- union(l1, l2)
  dat1 <- se1[genes,]
  dat2 <- se2[genes,]
  return(list("d1"=dat1, "d2"=dat2))
}

calcNetDiff <- function(dat, my_col){
  netyes <- cor(t(assay(dat)[, dat[[my_col]] == "Y"]))
  netno  <- cor(t(assay(dat)[, dat[[my_col]] =="N"]))
  return(netyes-netno)
} # //TODO - change to general y/n


# matrices --> edge lists
mat2elist <- function(netdiff, cor.cut=0.5){
  nrows=nrow(netdiff)
  cormat2 <- rep(1:nrows, each=nrows)
  cormat1 <- rep(1:nrows,nrows)
  el <- cbind(cormat1, cormat2, c(netdiff))
  melted <- melt(upper.tri(netdiff))
  melted <- melted[which(melted$value),]
  values <- netdiff[which(upper.tri(netdiff))]
  melted <- cbind(melted[,1:2], values)
  genes <- row.names(netdiff)
  melted[,1] <- genes[melted[,1]]
  melted[,2] <- genes[melted[,2]]
  row.names(melted) <- paste(melted[,1], melted[,2], sep="_")
  tosub <- melted
  tosel <- row.names(tosub[which(abs(tosub[,3])>cor.cut),])
  return(tosel)
}



```


```{r}

deEdges <- function(se, corsub, my_col){
  # get 50 most DE edges -- limma on EDGES
  group <- factor(se[[my_col]]) 
  design <- model.matrix(~0+group)
  cont.matrix <- makeContrasts(yesvsno = (groupY - groupN), levels = design)   
  fit <- lmFit(corsub, design)
  fit2 <- contrasts.fit(fit, cont.matrix)
  fit2e <- eBayes(fit2)
  toptable <- topTable(fit2e, number=nrow(corsub), adjust="fdr")
  return(toptable)
}


deGenes <- function(se, edges, my_col){
  #topgeneslist <-unique(unlist(lapply(rownames(edges)[1:nrow()], function(x) str_split(x, "_")[[1]])) )
  group <- factor(se[[my_col]]) 
  design <- model.matrix(~0+group)
  cont.matrix <- makeContrasts(yesvsno = (groupY - groupN), levels = design)    # TODO generalize
  fit <- lmFit(assay(se), design)
  fit2 <- contrasts.fit(fit, cont.matrix)
  fit2e <- eBayes(fit2)
  topDE <- topTable(fit2e, number=nrow(assay(se)), adjust="fdr")
  #topDE <- topDE[which(row.names(topDE) %in% topgeneslist),]
  #topgenesDE <- tibble("gene"=row.names(topDE), "t"=topDE$t)
  return(topDE)
}

# plot!!!
plotLIONESS <- function(edges, genes){
  edges2 <- t(matrix(unlist(c(strsplit(row.names(edges), "_"))),2))
  z <- cbind(edges2[1:50,], edges$logFC[1:50])
  g <- graph.data.frame(z, directed=FALSE)
  E(g)$weight <- as.numeric(z[,3])
  E(g)$color[E(g)$weight<0] <- "blue"
  E(g)$color[E(g)$weight>0] <- "red"
  E(g)$weight <- 1
  
  # coloring
  nodeorder <- cbind(V(g)$name, 1:length(V(g)))
  nodes <- merge(nodeorder, genes, by.x=1, by.y=1)
  nodes <- nodes[order(as.numeric(as.character(nodes[,2]))),]
  nodes[,3] <- as.numeric(as.character(nodes[,3]))
  nodes <- nodes[,-2]
  V(g)$weight <- nodes[,2]
  
  # make a color palette
  mypalette4 <- colorRampPalette(c("blue","white","white","red"), space="Lab")(256) 
  breaks2a <- seq(min(V(g)$weight), 0, length.out=128)
  breaks2b <- seq(0.00001, max(V(g)$weight)+0.1,length.out=128)
  breaks4 <- c(breaks2a,breaks2b)
  
  # select bins for colors
  bincol <- rep(NA, length(V(g)))
  for(i in 1:length(V(g))){
      bincol[i] <- min(which(breaks4>V(g)$weight[i]))
  }
  bincol <- mypalette4[bincol]
  
  # add colors to nodes
  V(g)$color <- bincol
  
  par(mar=c(0,0,0,0))
  plot(g, vertex.label.cex=0.7, vertex.size=10, vertex.label.color = "black", vertex.label.font=3, edge.width=10*(abs(as.numeric(z[,3]))-0.7), vertex.color=V(g)$color)
}

```

```{r}

f_se <- makeSumExp(f_dat, f_meta %>% dplyr::select(geo_accession, is_smoker))
m_se <- makeSumExp(m_dat, m_meta %>% dplyr::select(geo_accession, is_smoker))

s_se <- makeSumExp(s_dat, s_meta %>% dplyr::select(geo_accession, is_male))
ns_se <- makeSumExp(ns_dat, ns_meta %>% dplyr::select(geo_accession, is_male))


f_se2 <- varExp(f_se)
m_se2 <- varExp(m_se)
fm_vars <- varExpEither(f_se, m_se)
f_se2 <- fm_vars[[1]]
m_se2 <- fm_vars[[2]]


s_se2 <- varExp(s_se)
ns_se2 <- varExp(ns_se)

sn_vars <- varExpEither(s_se, ns_se)
s_se2 <- sn_vars[[1]]
ns_se2 <- sn_vars[[2]]


f_ndf <- calcNetDiff(f_se2, "is_smoker")
m_ndf <- calcNetDiff(m_se2, "is_smoker")

s_ndf <- calcNetDiff(s_se2, "is_male")
ns_ndf <- calcNetDiff(ns_se2, "is_male")

f_tosel <- mat2elist(f_ndf)
m_tosel <- mat2elist(m_ndf)
fm_tosel <- union(f_tosel, m_tosel)
s_tosel <- mat2elist(s_ndf)
ns_tosel <- mat2elist(ns_ndf)
sn_tosel <- union(s_tosel, ns_tosel)

# create coexp network

f_cormat <- lioness(f_se2, netFun)
f_corsub <- assay(f_cormat[which(row.names(f_cormat) %in% fm_tosel), ])
m_cormat <- lioness(m_se2, netFun)
m_corsub <- assay(m_cormat[which(row.names(m_cormat) %in% fm_tosel), ])

s_cormat <- lioness(s_se2, netFun)
s_corsub <- assay(s_cormat[which(row.names(s_cormat) %in% sn_tosel), ])
ns_cormat <- lioness(ns_se2, netFun)
ns_corsub <- assay(ns_cormat[which(row.names(ns_cormat) %in% sn_tosel), ])



f_edges <- deEdges(f_se, f_corsub, "is_smoker")
m_edges <- deEdges(m_se, m_corsub, "is_smoker")

s_edges <- deEdges(s_se, s_corsub, "is_male")
ns_edges <- deEdges(ns_se, ns_corsub, "is_male")

f_genes <- deGenes(f_se2, f_edges, "is_smoker")
m_genes <- deGenes(m_se2, m_edges, "is_smoker")

f_genes$gene <- rownames(f_genes)
m_genes$gene <- rownames(m_genes)

comb_genes <- f_genes %>% full_join(m_genes, by=c("gene"))
sig_genes <- comb_genes %>% 
  dplyr::select(-contains("AveExpr"), -contains("P.Value"), -contains("B")) %>% 
  filter(adj.P.Val.x < 0.05 | adj.P.Val.y < 0.05) %>%
  as_tibble()

opp_sig <- sig_genes %>% filter(abs((logFC.x - logFC.y)/logFC.x )> 2)

ggplot(comb_genes, aes(x=logFC.x, y=logFC.y))+
         geom_point(alpha=0.3)+
         geom_point(data=sig_genes, col="red", alpha=0.5)+

  geom_vline(xintercept=0, col="gray", alpha=0.8)+
  geom_hline(yintercept=0, col="gray", alpha=0.8)+
  theme_bw()+
  xlab("log(S-NS) in females")+
  ylab("log(S-NS) in males")+
           geom_label_repel(data=sig_genes %>% 
                            filter(abs(logFC.x) > 10 |
                                     abs(logFC.y) > 10), 
                          aes(label=gene), size=2)


ggplot(comb_genes, aes(x=logFC.x, y=logFC.y))+
         geom_point(alpha=0.3)+
         geom_point(data=sig_genes, col="red", alpha=0.5)+
         geom_point(data=opp_sig, col="blue", alpha=0.5)+

  geom_vline(xintercept=0, col="gray", alpha=0.8)+
  geom_hline(yintercept=0, col="gray", alpha=0.8)+
  theme_bw()+
  xlab("log(S-NS) in females")+
  ylab("log(S-NS) in males")+
           geom_label_repel(data=sig_genes %>% 
                            filter(abs(logFC.x) > 15 |
                                     abs(logFC.y) > 15), 
                          aes(label=gene), size=2)+
  geom_label_repel(data=opp_sig, col="purple", 
                          aes(label=gene), size=2)

dat3 <- (gse11784.3 %>% filter(hgnc_symbol %in% sig_genes$gene))
dat4 <- dat3[,study1.3$geo_accession]
rownames(dat4) <- dat3$hgnc_symbol
design <- model.matrix(~0+is_male+is_smoker+is_male*is_smoker, data=study1.3)
fit <- lmFit(dat4, design=design)
fit2e <- eBayes(fit)
topDE <- topTable(fit2e, coef="is_maleY:is_smokerY", number=nrow(dat3), adjust="fdr")
head(topDE)
# go enrichment - is it different?


s_genes <- deGenes(s_se2, s_edges, "is_male")
ns_genes <- deGenes(ns_se2, ns_edges, "is_male")
s_genes$gene <- rownames(s_genes)
ns_genes$gene <- rownames(ns_genes)

s_genes_sig <- s_genes %>% filter(adj.P.Val < 0.1)
ns_genes_sig <- ns_genes %>% filter(adj.P.Val < 0.1)
setdiff(s_genes_sig$gene, ns_genes_sig$gene)
# HSPA8??



#plotLIONESS(f_edges, f_genes)  
#plotLIONESS(m_edges, m_genes)  


#plotLIONESS(s_edges, s_genes)  
#plotLIONESS(ns_edges, ns_genes)  
```

Look at the genes
```{r}
head(s_genes)
head(ns_genes)


reformatEdge <- function(x) {
  y <- str_split(x, "_")[[1]];
  z <- y[order(y)];
  return(paste(z, collapse="_"))
}



sns_genes <- full_join(
  s_genes %>% dplyr::rename(t.s=t), 
  ns_genes %>% dplyr::rename(t.ns=t), by=c("gene"))

s_edges2 <- s_edges %>% dplyr::select(logFC) 
s_edges2$edge <- rownames(s_edges)
s_edges2$edge <- sapply(s_edges2$edge, reformatEdge)

ns_edges2 <- ns_edges %>% dplyr::select(logFC) 
ns_edges2$edge <- rownames(ns_edges)
ns_edges2$edge <- sapply(ns_edges2$edge, reformatEdge)

sns_edges <- full_join(
 s_edges2 %>% dplyr::rename(logFC.s=logFC), 
  ns_edges2 %>% dplyr::rename(logFC.ns=logFC), by=c("edge"))

sns_edges2 <- sns_edges %>% separate(edge, into=c("e1", "e2"), sep="_") %>% dplyr::select(e1, e2, everything())

sns_edges3 <- bind_rows(sns_edges2 %>% arrange(desc(abs(logFC.s))) %>% head(25), sns_edges2 %>% arrange(desc(abs(logFC.ns))) %>% head(25)) %>% unique()


fm_genes <- full_join(
  f_genes %>% dplyr::rename(t.f=t), 
  m_genes %>% dplyr::rename(t.m=t), by=c("gene"))

f_edges2 <- f_edges %>% dplyr::select(logFC) 
f_edges2$edge <- rownames(f_edges)
f_edges2$edge <- sapply(f_edges2$edge, reformatEdge)

m_edges2 <- m_edges %>% dplyr::select(logFC) 
m_edges2$edge <- rownames(m_edges)
m_edges2$edge <- sapply(m_edges2$edge, reformatEdge)

fm_edges <- full_join(
 f_edges2 %>% dplyr::rename(logFC.f=logFC), 
m_edges2 %>% dplyr::rename(logFC.m=logFC), by=c("edge"))

fm_edges2 <- fm_edges %>% separate(edge, into=c("e1", "e2"), sep="_") %>% dplyr::select(e1, e2, everything())

fm_edges3 <- bind_rows(fm_edges2 %>% arrange(desc(abs(logFC.f))) %>% head(25), fm_edges2 %>% arrange(desc(abs(logFC.m))) %>% head(25)) %>% unique()
```


what can we do with the genes/edges?
```{r}
plotG <- function(genes, g, g_weight, my_col="t.s"){
  E(g)$weight <- as.numeric(g_weight)
  E(g)$color[E(g)$weight<0] <- "blue"
  E(g)$color[E(g)$weight>0] <- "red"
  E(g)$weight <- 1
  
  # coloring
  nodeorder <- tibble("gene"=V(g)$name, "order"=1:length(V(g)))
  nodes <- left_join(nodeorder, genes) %>%
    replace_na(replace=list("t.s"=0, "t.ns"=0))
  V(g)$weight <- as.numeric(nodes[[my_col]])

  # make a color palette
  mypalette4 <- colorRampPalette(c("blue","white","white","red"), space="Lab")(256) 
  breaks2a <- seq(min(V(g)$weight), 0, length.out=128)
  breaks2b <- seq(0.00001, max(V(g)$weight)+0.1,length.out=128)
  breaks4 <- c(breaks2a,breaks2b)
  
  # select bins for colors
  bincol <- rep(NA, length(V(g)))
  for(i in 1:length(V(g))){
      bincol[i] <- min(which(breaks4>V(g)$weight[i]))
  }
  bincol <- mypalette4[bincol]
  V(g)$weight <- 1

  # add colors to nodes
  V(g)$color <- bincol
  par(mar=c(0,0,0,0))
  plot(g, vertex.label.cex=0.7, vertex.size=10, 
       vertex.label.color = "black", vertex.label.font=3,
       edge.width=5*(abs(as.numeric(g_weight))-0.9))
}



g1 <- graph.data.frame(sns_edges3 %>% dplyr::select(e1, e2), directed=FALSE)
set.seed(1)
plotG(sns_genes, g1, sns_edges3$logFC.s, my_col="t.s")
set.seed(1)
plotG(sns_genes, g1, sns_edges3$logFC.ns, my_col="t.ns")
g2 <- graph.data.frame(fm_edges3 %>% dplyr::select(e1, e2), directed=FALSE)
set.seed(1)
plotG(fm_genes, g2, fm_edges3$logFC.f, my_col="t.f")
set.seed(1)
plotG(fm_genes, g2, fm_edges3$logFC.m, my_col="t.m")
library(ggrepel)
ggplot(fm_edges2, 
       aes(x=logFC.f, y=logFC.m))+
  geom_point(alpha=0.3)+theme_bw()+
  geom_vline(xintercept=0, col="gray", alpha=0.8)+
  geom_hline(yintercept=0, col="gray", alpha=0.8)+
  geom_point(data=fm_edges3, col="red")+
  #geom_label_repel(data=fm_edges3 %>% unite(col="edge", c(e1, e2), sep="-"),size=2, aes(label=edge))+
  xlab("log(S-NS) in females")+
  ylab("log(S-NS) in males")

```

```{r}
# --- r.jive --- #
# PROBLEM: two classes not interaction  
#devtools::install_github("https://github.com/cran/SpatioTemporal")
#devtools::install_github("https://github.com/cran/r.jive")
library(r.jive) # --> joint, individual, noise
res <- jive(list(f_smokers, m_smokers)) # could try m vs f smokers? note -- how slow would this be? do we need to filter? (probs)

# ----  tSNR ---- #

```
